🔧 Move file hash/copy functions from myst-cli to myst-cli-utils package
