+++ packages/myst-cli-utils/src/index.ts
@@ -11,4 +11,4 @@ export { exec, makeExecutable } from './exec';
 export { clirun, tic } from './utils';
 export { isUrl } from './isUrl';
 export { Session, getSession } from './session';
-export { writeFileToFolder } from './filesystem';
+export { computeHash, hashAndCopyStaticFile, writeFileToFolder } from './filesystem';
