+++ packages/myst-cli/src/process/site.ts
@@ -3,7 +3,7 @@ import yaml from 'js-yaml';
 import { basename, extname, join } from 'path';
 import chalk from 'chalk';
 import { Inventory, Domains } from 'intersphinx';
-import { writeFileToFolder, tic } from 'myst-cli-utils';
+import { writeFileToFolder, tic, hashAndCopyStaticFile } from 'myst-cli-utils';
 import { toText } from 'myst-common';
 import type { SiteProject } from 'myst-config';
 import type { Export } from 'myst-frontmatter';
@@ -21,7 +21,6 @@ import { castSession } from '../session';
 import type { ISession } from '../session/types';
 import { watch, selectors } from '../store';
 import { transformWebp } from '../transforms';
-import { hashAndCopyStaticFile } from '../utils';
 import { combineProjectCitationRenderers } from './citations';
 import { loadIntersphinx } from './intersphinx';
 import type { PageReferenceStates, TransformFn } from './mdast';
