--- packages/myst-cli/src/utils/computeHash.ts
@@ -1,5 +1,0 @@
-import { createHash } from 'crypto';
-
-export function computeHash(content: string) {
-  return createHash('md5').update(content).digest('hex');
-}
