+++ packages/myst-cli/src/utils/index.ts
@@ -1,11 +1,9 @@
 export * from './addWarningForFile';
 export * from './check';
-export * from './computeHash';
 export * from './createTempFolder';
 export * from './fileInfo';
 export * from './filterFilenamesByExtension';
 export * from './getAllBibtexFiles';
-export * from './hashAndCopyStaticFile';
 export * from './isDirectory';
 export * from './logMessagesFromVFile';
 export * from './nextLevel';
